requirements.txt
PyHyphen==2.0.4
PyYAML==3.10
ipython==1.1.0
nltk==2.0.4
numpy==1.8.0
tweepy==2.1
wsgiref==0.1.2
